+++ tests/TestCase.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Butler\Service\Tests;
 
-use Illuminate\Auth\GenericUser;
 use Illuminate\Contracts\Console\Kernel;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
@@ -18,20 +17,5 @@ abstract class TestCase extends BaseTestCase
         $app->make(Kernel::class)->bootstrap();
 
         return $app;
-    }
-
-    protected function actingAsUser(array $data = [])
-    {
-        $user = new GenericUser(array_merge([
-            'id' => rand(0, 999),
-            'username' => 'username',
-            'name' => 'name',
-            'email' => 'user@example.com',
-            'oauth_token' => 'token',
-            'oauth_refresh_token' => 'refresh-token',
-            'remember_token' => null,
-        ], $data));
-
-        return $this->actingAs($user, 'web');
     }
 }